feat: complete cjs to mjs task
